ChannelVideoをActivityに
