+++ app/src/main/kotlin/com/numero/itube/di/ApplicationComponent.kt
@@ -1,9 +1,9 @@
 package com.numero.itube.di
 
+import com.numero.itube.activity.ChannelDetailActivity
 import com.numero.itube.activity.PlayerActivity
 import com.numero.itube.activity.SearchActivity
 import com.numero.itube.activity.TopActivity
-import com.numero.itube.fragment.ChannelVideoListFragment
 import dagger.Component
 import javax.inject.Singleton
 
@@ -15,9 +15,8 @@ import javax.inject.Singleton
     (RepositoryModule::class)
 ])
 interface ApplicationComponent {
-    fun inject(channelVideoListFragment: ChannelVideoListFragment)
-
     fun inject(topActivity: TopActivity)
     fun inject(searchActivity: SearchActivity)
     fun inject(playerActivity: PlayerActivity)
+    fun inject(channelDetailActivity: ChannelDetailActivity)
 }
